WIP Restore tests building
